Get diagnostics from Boogie
